2024.1.25

+++ back-end/model/mall/response/mall_goods_info.go
@@ -1,11 +1,11 @@
 package response
 
 type GoodsSearchResponse struct {
-	GoodsId       int    `json:"goodsId"`
-	GoodsName     string `json:"goodsName"`
-	GoodsIntro    string `json:"goodsIntro"`
-	GoodsCoverImg string `json:"goodsCoverImg"`
-	SellingPrice  int    `json:"sellingPrice"`
+	GoodsId       int     `json:"goodsId"`
+	GoodsName     string  `json:"goodsName"`
+	GoodsIntro    string  `json:"goodsIntro"`
+	GoodsCoverImg string  `json:"goodsCoverImg"`
+	SellingPrice  float64 `json:"sellingPrice"`
 }
 
 type GoodsInfoDetailResponse struct {
@@ -13,7 +13,7 @@ type GoodsInfoDetailResponse struct {
 	GoodsName          string   `json:"goodsName"`
 	GoodsIntro         string   `json:"goodsIntro"`
 	GoodsCoverImg      string   `json:"goodsCoverImg"`
-	SellingPrice       int      `json:"sellingPrice"`
+	SellingPrice       float64  `json:"sellingPrice"`
 	GoodsDetailContent string   `json:"goodsDetailContent"  `
 	OriginalPrice      int      `json:"originalPrice" `
 	Tag                string   `json:"tag" form:"tag" `
