+++ back-end/model/mall/response/mall_index_config.go
@@ -1,10 +1,10 @@
 package response
 
 type MallIndexConfigGoodsResponse struct {
-	GoodsId       int    `json:"goodsId"`
-	GoodsName     string `json:"goodsName"`
-	GoodsIntro    string `json:"goodsIntro"`
-	GoodsCoverImg string `json:"goodsCoverImg"`
-	SellingPrice  int    `json:"sellingPrice"`
-	Tag           string `json:"tag"`
+	GoodsId       int     `json:"goodsId"`
+	GoodsName     string  `json:"goodsName"`
+	GoodsIntro    string  `json:"goodsIntro"`
+	GoodsCoverImg string  `json:"goodsCoverImg"`
+	SellingPrice  float64 `json:"sellingPrice"`
+	Tag           string  `json:"tag"`
 }
