+++ back-end/model/mall/response/mall_recommend.go
@@ -1,8 +1,8 @@
 package response
 
 type RecommendResponse struct {
-	GoodsId       int    `json:"goodsId"`
-	GoodsName     string `json:"goodsName"`
-	GoodsCoverImg string `json:"goodsCoverImg"`
-	SellingPrice  int    `json:"sellingPrice"`
+	GoodsId       int     `json:"goodsId"`
+	GoodsName     string  `json:"goodsName"`
+	GoodsCoverImg string  `json:"goodsCoverImg"`
+	SellingPrice  float64 `json:"sellingPrice"`
 }
